refactor: 제품 필터링 조건 Context 리팩토링
