+++ src/constants/ProductsFilterinitial.tsx
@@ -1,0 +1,5 @@
+export const initialState = {
+  isFilterSale: null,
+  isFilterExclusive: null,
+  isFilterSoldOut: null,
+};
