+++ src/contexts/ProductsFilter/action.ts
@@ -1,0 +1,5 @@
+import * as TYPE from './types';
+
+export const getProductFilterAction = (actionIncome: TYPE.ActionIncomeType) => {
+  return { type: actionIncome };
+};
