+++ src/contexts/ProductsFilter/types.ts
@@ -1,0 +1,17 @@
+export type ProductsFilterStateType = {
+  isFilterSale: boolean | null;
+  isFilterExclusive: boolean | null;
+  isFilterSoldOut: boolean | null;
+};
+
+export type ProductsFliterAction =
+  | { type: 'CLICKED_SALE' }
+  | { type: 'CLICKED_EXCLUSIVE' }
+  | { type: 'CLICKED_SOLD_OUT' }
+  | { type: 'RESET' };
+
+export type ActionIncomeType =
+  | 'CLICKED_SALE'
+  | 'CLICKED_EXCLUSIVE'
+  | 'CLICKED_SOLD_OUT'
+  | 'RESET';
